PatchLLMAPI done
