+++ wandb/sdk/lib/timer.py
@@ -4,8 +4,8 @@ from typing import Any
 
 class Timer:
     def __init__(self) -> None:
-        self.start: float = time.perf_counter()
         self.start_time: float = time.time()
+        self.start: float = time.perf_counter()
         self.stop: float = self.start
 
     def __enter__(self) -> "Timer":
